+++ src/queries/cart/index.ts
@@ -2,5 +2,7 @@
  * @file Automatically generated by barrelsby.
  */
 
+export * from './useAddProductToCartMutation';
 export * from './useCartsQuery';
 export * from './useInitializeCarts';
+export * from './useRemoveProductToCartMutation';
