+++ src/queries/product/index.ts
@@ -2,6 +2,4 @@
  * @file Automatically generated by barrelsby.
  */
 
-export * from './useAddProductToCartMutation';
 export * from './useProductsQuery';
-export * from './useRemoveProductToCartMutation';
